+++ config/environment.rb
@@ -16,7 +16,7 @@ UNICODE_NORMALIZATION_FORM = :c
 # Default user preferences.
 DEFAULT_USER_PREFERENCES = { :graph_format => "png", :graph_method => "unsorted" }
 
-RAILS_GEM_VERSION = '2.3.11' unless defined? RAILS_GEM_VERSION
+RAILS_GEM_VERSION = '2.3.12' unless defined? RAILS_GEM_VERSION
 
 # Bootstrap the Rails environment, frameworks, and default configuration
 require File.join(File.dirname(__FILE__), 'boot')
@@ -55,6 +55,7 @@ Rails::Initializer.run do |config|
   # Make Active Record use UTC-base instead of local time
   # config.active_record.default_timezone = :utc
 
+  config.gem 'mysql2', :version => '< 0.3'
   config.gem 'warden', :version => '~> 0.9'
   config.gem 'devise', :version => '~> 1.0.8'
   config.gem 'will_paginate', :version => '~> 2.3.2'